+++ cmd/root.go
@@ -36,7 +36,7 @@ var rootCmd = &cobra.Command{
 	Use:           "safira",
 	Short:         "Safira is a toolkit for Vertigo iPaaS",
 	Long:          "Safira is a toolkit for Vertigo iPaaS",
-	Version:       "v0.0.15",
+	Version:       "v0.0.16",
 	SilenceUsage:  true,
 	SilenceErrors: true,
 }